+++ core/src/codecs/DeletedDocsFormat.cpp
@@ -29,6 +29,7 @@
 #include "db/Utils.h"
 #include "utils/Exception.h"
 #include "utils/Log.h"
+#include "utils/TimeRecorder.h"
 
 namespace milvus {
 namespace codec {
@@ -45,23 +46,29 @@ Status
 DeletedDocsFormat::Read(const storage::FSHandlerPtr& fs_ptr, const std::string& file_path,
                         segment::DeletedDocsPtr& deleted_docs) {
     const std::string full_file_path = file_path + DELETED_DOCS_POSTFIX;
-
+    milvus::TimeRecorderAuto recorder("DeletedDocsFormat::Read:" + full_file_path);
     if (!fs_ptr->reader_ptr_->Open(full_file_path)) {
         return Status(SERVER_CANNOT_OPEN_FILE, "Fail to open deleted docs file: " + full_file_path);
     }
     CHECK_MAGIC_VALID(fs_ptr);
-    CHECK_SUM_VALID(fs_ptr);
+    std::vector<char> header;
+    header.resize(HEADER_SIZE);
+    fs_ptr->reader_ptr_->Read(header.data(), HEADER_SIZE);
 
-    HeaderMap map = ReadHeaderValues(fs_ptr);
+    HeaderMap map = TransformHeaderData(header);
     size_t num_bytes = stol(map.at("size"));
 
     auto deleted_docs_size = num_bytes / sizeof(engine::offset_t);
     std::vector<engine::offset_t> deleted_docs_list;
     deleted_docs_list.resize(deleted_docs_size);
 
-    fs_ptr->reader_ptr_->Seekg(MAGIC_SIZE + HEADER_SIZE);
     fs_ptr->reader_ptr_->Read(deleted_docs_list.data(), num_bytes);
+
+    uint32_t record;
+    fs_ptr->reader_ptr_->Read(&record, SUM_SIZE);
     fs_ptr->reader_ptr_->Close();
+
+    CHECK_SUM_VALID(header.data(), reinterpret_cast<const char*>(deleted_docs_list.data()), num_bytes, record);
 
     deleted_docs = std::make_shared<segment::DeletedDocs>(deleted_docs_list);
 
@@ -72,7 +79,7 @@ Status
 DeletedDocsFormat::Write(const storage::FSHandlerPtr& fs_ptr, const std::string& file_path,
                          const segment::DeletedDocsPtr& deleted_docs) {
     const std::string full_file_path = file_path + DELETED_DOCS_POSTFIX;
-
+    milvus::TimeRecorderAuto recorder("DeletedDocsFormat::Write:" + full_file_path);
     auto deleted_docs_list = deleted_docs->GetDeletedDocs();
     size_t num_bytes = sizeof(engine::offset_t) * deleted_docs->GetCount();
 
@@ -111,13 +118,27 @@ DeletedDocsFormat::ReadSize(const storage::FSHandlerPtr& fs_ptr, const std::stri
         return Status(SERVER_CANNOT_CREATE_FILE, "Fail to open deleted docs file: " + full_file_path);
     }
     CHECK_MAGIC_VALID(fs_ptr);
-    CHECK_SUM_VALID(fs_ptr);
 
-    HeaderMap map = ReadHeaderValues(fs_ptr);
+    std::vector<char> header;
+    header.resize(HEADER_SIZE);
+    fs_ptr->reader_ptr_->Read(header.data(), HEADER_SIZE);
+
+    HeaderMap map = TransformHeaderData(header);
     size_t num_bytes = stol(map.at("size"));
 
-    size = num_bytes / sizeof(engine::offset_t);
+    std::vector<char> data;
+
+    fs_ptr->reader_ptr_->Read(data.data(), num_bytes);
+
+    uint32_t record;
+    fs_ptr->reader_ptr_->Read(&record, SUM_SIZE);
     fs_ptr->reader_ptr_->Close();
+
+    CHECK_SUM_VALID(header.data(), reinterpret_cast<const char*>(data.data()), num_bytes, record);
+
+    fs_ptr->reader_ptr_->Close();
+
+    size = num_bytes / sizeof(engine::offset_t);
 
     return Status::OK();
 }