+++ core/src/db/insert/MemCollection.cpp
@@ -40,7 +40,7 @@ namespace engine {
 
 MemCollection::MemCollection(int64_t collection_id, const DBOptions& options)
     : collection_id_(collection_id), options_(options) {
-    GetSegmentRowCount(collection_id_, segment_row_count_);
+    GetSegmentRowLimit(collection_id_, segment_row_limit_);
 }
 
 Status
@@ -58,8 +58,8 @@ MemCollection::Add(int64_t partition_id, const DataChunkPtr& chunk, idx_t op_id)
     int64_t chunk_size = utils::GetSizeOfChunk(chunk);
 
     Status status;
-    if (current_mem_segment == nullptr || chunk->count_ >= segment_row_count_ ||
-        current_mem_segment->GetCurrentRowCount() >= segment_row_count_ ||
+    if (current_mem_segment == nullptr || chunk->count_ >= segment_row_limit_ ||
+        current_mem_segment->GetCurrentRowCount() >= segment_row_limit_ ||
         current_mem_segment->GetCurrentMem() + chunk_size > MAX_MEM_SEGMENT_SIZE) {
         MemSegmentPtr new_mem_segment = std::make_shared<MemSegment>(collection_id_, partition_id, options_);
         status = new_mem_segment->Add(chunk, op_id);
@@ -91,6 +91,9 @@ MemCollection::Delete(const std::vector<idx_t>& ids, idx_t op_id) {
     for (auto& id : ids) {
         ids_to_delete_.insert(id);
     }
+
+    // record max delete operation id here, for the case that no insert but only delete action performed
+    max_delete_op_id_ = (op_id > max_delete_op_id_) ? op_id : max_delete_op_id_;
 
     // Add the id to mem segments so it can be applied during the next flush
     std::lock_guard<std::mutex> lock(mem_mutex_);
@@ -337,12 +340,16 @@ MemCollection::CreateDeletedDocsBloomFilter(const std::shared_ptr<snapshot::Comp
 Status
 MemCollection::SerializeSegments() {
     std::lock_guard<std::mutex> lock(mem_mutex_);
-    if (mem_segments_.empty()) {
-        return Status::OK();
-    }
 
     snapshot::ScopedSnapshotT ss;
     STATUS_CHECK(snapshot::Snapshots::GetInstance().GetSnapshot(ss, collection_id_));
+
+    if (mem_segments_.empty()) {
+        // for the case that no insert but only delete action performed
+        // notify wal manager the delete operations has been done
+        WalManager::GetInstance().OperationDone(ss->GetName(), max_delete_op_id_);
+        return Status::OK();
+    }
 
     snapshot::OperationContext context;
     auto operation = std::make_shared<snapshot::MultiSegmentsOperation>(context, ss);