+++ core/src/segment/SegmentReader.cpp
@@ -26,6 +26,9 @@
 #include "db/Types.h"
 #include "db/Utils.h"
 #include "db/snapshot/ResourceHelper.h"
+#include "knowhere/index/structured_index/StructuredIndexSort.h"
+#include "knowhere/index/vector_index/IndexBinaryIDMAP.h"
+#include "knowhere/index/vector_index/IndexIDMAP.h"
 #include "knowhere/index/vector_index/VecIndex.h"
 #include "knowhere/index/vector_index/VecIndexFactory.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
@@ -37,6 +40,45 @@
 
 namespace milvus {
 namespace segment {
+namespace {
+template <typename T>
+knowhere::IndexPtr
+CreateSortedIndex(engine::BinaryDataPtr& raw_data) {
+    if (raw_data == nullptr) {
+        return nullptr;
+    }
+
+    auto count = raw_data->data_.size() / sizeof(T);
+    auto index_ptr =
+        std::make_shared<knowhere::StructuredIndexSort<T>>(count, reinterpret_cast<const T*>(raw_data->data_.data()));
+    return std::static_pointer_cast<knowhere::Index>(index_ptr);
+}
+
+Status
+CreateStructuredIndex(const engine::DataType field_type, engine::BinaryDataPtr& raw_data,
+                      knowhere::IndexPtr& index_ptr) {
+    switch (field_type) {
+        case engine::DataType::INT32: {
+            index_ptr = CreateSortedIndex<int32_t>(raw_data);
+            break;
+        }
+        case engine::DataType::INT64: {
+            index_ptr = CreateSortedIndex<int64_t>(raw_data);
+            break;
+        }
+        case engine::DataType::FLOAT: {
+            index_ptr = CreateSortedIndex<float>(raw_data);
+            break;
+        }
+        case engine::DataType::DOUBLE: {
+            index_ptr = CreateSortedIndex<double>(raw_data);
+            break;
+        }
+        default: { return Status(DB_ERROR, "Field is not structured type"); }
+    }
+    return Status::OK();
+}
+}  // namespace
 
 SegmentReader::SegmentReader(const std::string& dir_root, const engine::SegmentVisitorPtr& segment_visitor,
                              bool initialize)
@@ -98,7 +140,7 @@ SegmentReader::Load() {
     segment::DeletedDocsPtr deleted_docs_ptr;
     LoadDeletedDocs(deleted_docs_ptr);
 
-    STATUS_CHECK(LoadVectorIndice());
+    STATUS_CHECK(LoadIndice());
 
     return Status::OK();
 }
@@ -106,7 +148,7 @@ SegmentReader::Load() {
 Status
 SegmentReader::LoadField(const std::string& field_name, engine::BinaryDataPtr& raw, bool to_cache) {
     try {
-        TimeRecorder recorder("SegmentReader::LoadField: " + field_name);
+        TimeRecorderAuto recorder("SegmentReader::LoadField: " + field_name);
 
         segment_ptr_->GetFixedFieldData(field_name, raw);
         if (raw != nullptr) {
@@ -136,8 +178,6 @@ SegmentReader::LoadField(const std::string& field_name, engine::BinaryDataPtr& r
         }
 
         segment_ptr_->SetFixedFieldData(field_name, raw);
-
-        recorder.RecordSection("read " + file_path);
     } catch (std::exception& e) {
         std::string err_msg = "Failed to load raw vectors: " + std::string(e.what());
         LOG_ENGINE_ERROR_ << err_msg;
@@ -169,24 +209,84 @@ SegmentReader::LoadEntities(const std::string& field_name, const std::vector<int
     try {
         TimeRecorderAuto recorder("SegmentReader::LoadEntities: " + field_name);
 
-        auto field_visitor = segment_visitor_->GetFieldVisitor(field_name);
-        if (field_visitor == nullptr) {
-            return Status(DB_ERROR, "Invalid field_name");
-        }
-        auto raw_visitor = field_visitor->GetElementVisitor(engine::FieldElementType::FET_RAW);
-        std::string file_path =
-            engine::snapshot::GetResPath<engine::snapshot::SegmentFile>(dir_collections_, raw_visitor->GetFile());
-
         int64_t field_width = 0;
         STATUS_CHECK(segment_ptr_->GetFixedFieldWidth(field_name, field_width));
         if (field_width <= 0) {
             return Status(DB_ERROR, "Invalid field width");
         }
 
+        auto field_visitor = segment_visitor_->GetFieldVisitor(field_name);
+        if (field_visitor == nullptr) {
+            return Status(DB_ERROR, "Invalid field name");
+        }
+
+        // copy from cache function
+        auto copy_data = [&](uint8_t* src_data, int64_t src_data_size, const std::vector<int64_t>& offsets,
+                             int64_t field_width, engine::BinaryDataPtr& raw) -> Status {
+            if (src_data == nullptr) {
+                return Status(DB_ERROR, "src_data is null pointer");
+            }
+            int64_t total_bytes = offsets.size() * field_width;
+            raw = std::make_shared<engine::BinaryData>();
+            raw->data_.resize(total_bytes);
+
+            // copy from cache
+            int64_t target_poz = 0;
+            for (auto offset : offsets) {
+                int64_t src_poz = offset * field_width;
+                if (offset < 0 || src_poz + field_width > src_data_size) {
+                    return Status(DB_ERROR, "Invalid entity offset");
+                }
+
+                memcpy(raw->data_.data() + target_poz, src_data + src_poz, field_width);
+                target_poz += field_width;
+            }
+
+            return Status::OK();
+        };
+
+        // if raw data is alrady in cache, copy from cache
+        const engine::snapshot::FieldPtr& field = field_visitor->GetField();
+        engine::BinaryDataPtr field_data;
+        segment_ptr_->GetFixedFieldData(field_name, field_data);
+        if (field_data != nullptr) {
+            return copy_data(field_data->data_.data(), field_data->data_.size(), offsets, field_width, raw);
+        }
+
+        // for vector field, the LoadVectorIndex() could create a IDMAP index in cache, copy from the index
+        if (engine::IsVectorField(field)) {
+            std::string temp_index_path;
+            GetTempIndexPath(field->GetName(), temp_index_path);
+
+            uint8_t* src_data = nullptr;
+            int64_t src_data_size = 0;
+            if (auto data_obj = cache::CpuCacheMgr::GetInstance().GetItem(temp_index_path)) {
+                auto index_ptr = std::static_pointer_cast<knowhere::VecIndex>(data_obj);
+                if (index_ptr->index_type() == knowhere::IndexEnum::INDEX_FAISS_IDMAP) {
+                    auto idmap_index = std::static_pointer_cast<knowhere::IDMAP>(index_ptr);
+                    src_data = (uint8_t*)idmap_index->GetRawVectors();
+                    src_data_size = idmap_index->Count() * field_width;
+                } else if (index_ptr->index_type() == knowhere::IndexEnum::INDEX_FAISS_BIN_IDMAP) {
+                    auto idmap_index = std::static_pointer_cast<knowhere::BinaryIDMAP>(index_ptr);
+                    src_data = (uint8_t*)idmap_index->GetRawVectors();
+                    src_data_size = idmap_index->Count() * field_width;
+                }
+            }
+
+            if (src_data) {
+                return copy_data(src_data, src_data_size, offsets, field_width, raw);
+            }
+        }
+
+        // read from storage
         codec::ReadRanges ranges;
         for (auto offset : offsets) {
             ranges.push_back(codec::ReadRange(offset * field_width, field_width));
         }
+
+        auto raw_visitor = field_visitor->GetElementVisitor(engine::FieldElementType::FET_RAW);
+        std::string file_path =
+            engine::snapshot::GetResPath<engine::snapshot::SegmentFile>(dir_collections_, raw_visitor->GetFile());
         auto& ss_codec = codec::Codec::instance();
         STATUS_CHECK(ss_codec.GetBlockFormat()->Read(fs_ptr_, file_path, ranges, raw));
     } catch (std::exception& e) {
@@ -312,13 +412,14 @@ SegmentReader::LoadVectorIndex(const std::string& field_name, knowhere::VecIndex
         // if index not specified, or index file not created, return a temp index(IDMAP type)
         auto index_visitor = field_visitor->GetElementVisitor(engine::FieldElementType::FET_INDEX);
         if (flat || index_visitor == nullptr || index_visitor->GetFile() == nullptr) {
-            // if the data is in cache, no need to read file
             std::string temp_index_path;
             GetTempIndexPath(field_name, temp_index_path);
             auto data_obj = cache::CpuCacheMgr::GetInstance().GetItem(temp_index_path);
             if (data_obj != nullptr) {
+                // if the temp index is in cache, no need to create it
                 index_ptr = std::static_pointer_cast<knowhere::VecIndex>(data_obj);
                 segment_ptr_->SetVectorIndex(field_name, index_ptr);
+                recorder.RecordSection("get temp index from cache");
             } else {
                 auto& json = field->GetParams();
                 if (json.find(knowhere::meta::DIM) == json.end()) {
@@ -424,7 +525,7 @@ SegmentReader::LoadVectorIndex(const std::string& field_name, knowhere::VecIndex
 Status
 SegmentReader::LoadStructuredIndex(const std::string& field_name, knowhere::IndexPtr& index_ptr) {
     try {
-        TimeRecorderAuto recorder("SegmentReader::LoadStructuredIndex");
+        TimeRecorder recorder("SegmentReader::LoadStructuredIndex: " + field_name);
 
         segment_ptr_->GetStructuredIndex(field_name, index_ptr);
         if (index_ptr != nullptr) {
@@ -442,9 +543,9 @@ SegmentReader::LoadStructuredIndex(const std::string& field_name, knowhere::Inde
             return Status(DB_ERROR, "Field is not structured type");
         }
 
-        // read field index
         auto index_visitor = field_visitor->GetElementVisitor(engine::FieldElementType::FET_INDEX);
         if (index_visitor && index_visitor->GetFile() != nullptr) {
+            // read field index
             std::string file_path =
                 engine::snapshot::GetResPath<engine::snapshot::SegmentFile>(dir_collections_, index_visitor->GetFile());
 
@@ -453,12 +554,36 @@ SegmentReader::LoadStructuredIndex(const std::string& field_name, knowhere::Inde
             if (data_obj == nullptr) {
                 STATUS_CHECK(ss_codec.GetStructuredIndexFormat()->Read(fs_ptr_, file_path, index_ptr));
                 cache::CpuCacheMgr::GetInstance().InsertItem(file_path, index_ptr);  // put into cache
+                recorder.RecordSection("read from storage");
             } else {
                 index_ptr = std::static_pointer_cast<knowhere::Index>(data_obj);
+                recorder.RecordSection("get from cache");
             }
+        } else {
+            // if index not specified, or index file not created, return a temp index(SORTED type)
+            std::string temp_index_path;
+            GetTempIndexPath(field_name, temp_index_path);
+            auto data_obj = cache::CpuCacheMgr::GetInstance().GetItem(temp_index_path);
+            if (data_obj != nullptr) {
+                // if the temp index is in cache, no need to create it
+                index_ptr = std::static_pointer_cast<knowhere::VecIndex>(data_obj);
+                recorder.RecordSection("get temp index from cache");
+            } else {
+                // create temp index and put into cache
+                engine::DataType field_type = engine::DataType::NONE;
+                STATUS_CHECK(segment_ptr_->GetFieldType(field_name, field_type));
 
-            segment_ptr_->SetStructuredIndex(field_name, index_ptr);
+                engine::BinaryDataPtr raw_data;
+                LoadField(field_name, raw_data, false);
+                STATUS_CHECK(CreateStructuredIndex(field_type, raw_data, index_ptr));
+
+                cache::CpuCacheMgr::GetInstance().InsertItem(temp_index_path, index_ptr);  // put into cache
+
+                recorder.RecordSection("create temp index");
+            }
         }
+
+        segment_ptr_->SetStructuredIndex(field_name, index_ptr);
     } catch (std::exception& e) {
         std::string err_msg = "Failed to load vector index: " + std::string(e.what());
         LOG_ENGINE_ERROR_ << err_msg;
@@ -469,7 +594,7 @@ SegmentReader::LoadStructuredIndex(const std::string& field_name, knowhere::Inde
 }
 
 Status
-SegmentReader::LoadVectorIndice() {
+SegmentReader::LoadIndice() {
     auto& field_visitors_map = segment_visitor_->GetFieldVisitors();
     for (auto& iter : field_visitors_map) {
         const engine::snapshot::FieldPtr& field = iter.second->GetField();
@@ -480,8 +605,6 @@ SegmentReader::LoadVectorIndice() {
             continue;
         }
 
-        std::string file_path =
-            engine::snapshot::GetResPath<engine::snapshot::SegmentFile>(dir_collections_, element_visitor->GetFile());
         if (engine::IsVectorField(field)) {
             knowhere::VecIndexPtr index_ptr;
             STATUS_CHECK(LoadVectorIndex(name, index_ptr));
@@ -635,7 +758,7 @@ SegmentReader::GetTempIndexPath(const std::string& field_name, std::string& path
     auto segment = segment_visitor_->GetSegment();
     path = engine::snapshot::GetResPath<engine::snapshot::Segment>(dir_collections_, segment);
     path += "/";
-    std::string temp_index_name = field_name + ".idmap";
+    std::string temp_index_name = field_name + ".tmp.index";
     path += temp_index_name;
 
     return Status::OK();